+++ DTBaseApp/src/Renderer/DX11Buffers.h
@@ -6,9 +6,8 @@ namespace DT
 	class VertexBuffer
 	{
 	public:
-		VertexBuffer(const void* vertices, uint32 size);
+		VertexBuffer(const void* vertices, uint64 size);
 		void Bind(uint32 stride);
-
 	private:
 		Microsoft::WRL::ComPtr<ID3D11Buffer> m_VertexBuffer;
 	};
